+++ src/NzbDrone.Core.Test/MusicTests/AlbumRepositoryTests/AlbumRepositoryFixture.cs
@@ -23,13 +23,6 @@ namespace NzbDrone.Core.Test.MusicTests.BookRepositoryTests
         [SetUp]
         public void Setup()
         {
-            AssertionOptions.AssertEquivalencyUsing(options =>
-            {
-                options.Using<DateTime>(ctx => ctx.Subject.Should().BeCloseTo(ctx.Expectation.ToUniversalTime())).WhenTypeIs<DateTime>();
-                options.Using<DateTime?>(ctx => ctx.Subject.Should().BeCloseTo(ctx.Expectation.Value.ToUniversalTime())).WhenTypeIs<DateTime?>();
-                return options;
-            });
-
             _author = new Author
             {
                 Name = "Alien Ant Farm",
@@ -164,8 +157,11 @@ namespace NzbDrone.Core.Test.MusicTests.BookRepositoryTests
             result.Should().BeEquivalentTo(_books.Skip(2).Take(1), BookComparerOptions);
         }
 
-        private EquivalencyOptions<Book> BookComparerOptions(EquivalencyOptions<Book> opts) => opts.ComparingByMembers<Book>()
-                .Excluding(ctx => ctx.SelectedMemberInfo.MemberType.IsGenericType && ctx.SelectedMemberInfo.MemberType.GetGenericTypeDefinition() == typeof(LazyLoaded<>))
+        private EquivalencyOptions<Book> BookComparerOptions(EquivalencyOptions<Book> opts) => opts
+                .Using<DateTime>(ctx => ctx.Subject.Should().BeCloseTo(ctx.Expectation.ToUniversalTime(), TimeSpan.FromMilliseconds(50))).WhenTypeIs<DateTime>()
+                .Using<DateTime?>(ctx => ctx.Subject.Should().BeCloseTo(ctx.Expectation.Value.ToUniversalTime(), TimeSpan.FromMilliseconds(50))).WhenTypeIs<DateTime?>()
+                .ComparingByMembers<Book>()
+                .Excluding(member => member.Type.IsGenericType && member.Type.GetGenericTypeDefinition() == typeof(LazyLoaded<>))
                 .Excluding(x => x.AuthorId)
                 .Excluding(x => x.ForeignEditionId);
     }